Rewrite list parsing, pre-parse proc params.
This replaces the list_length/list_at APIs, which are basically designed
to encourage O(n**2) time complexity, with a single-pass parser that
stores the results in a Vec. Which is where the results often wind up
going anyway.

I've used this to pre-parse the param list given to proc, rather than
doing it on every proc invocation. This speeds up calling an empty proc
with three args (the call-proc bench) by ~12%.