Change flatten_string semantics for 20% boost.
I was trying to remember which places in the code reused buffers, and in
the tokenizer/evaluator, a comment on strpieces made it very clear that
I intended it to be reused.

And yet:

    command.push(flatten_string(mem::take(&mut strpieces)));

That hands the allocation into flatten_string, forcing a new allocation
next time we push to strpieces. That is _not_ what I intended, and looks
like an overzealous search-replace when I started using mem::take more
aggressively.

Fixed that. Unexpectedly, this boosts benchmarks by 19-20% across the
board. What.