proc: fix argument handling bugs.
The argument loop in the cmd_proc closure was validating args length as
it went. In addition to being more work than is strictly necessary, this
introduced a bug: if it early-exited finding the arg count incorrect,
the interpreter would be left in the partially constructed function
scope! (This is because it mutated the scope _before_ checking the
arguments.)

The reason this bug could be triggered in the first place: user procs
were being registered with arity 0, meaning they can be called with any
number of arguments. I got this from partcl, but I've decided it's
wrong. partcl implicitly binds any forgotten arguments to the empty
string, which is why it used arity 0, but this is _not_ what Tcl does,
and would be confusing for me as a user. wartcl has never behaved this
way, making the use of arity 0 silly.

So, this now binds the closure to its specific arity, and moves argument
validation out of the loop.