tokenizer: return roles, make fewer passes
Originally, wartcl (like partcl) would detect things like variable
references, and return them as bare strings. They'd then be passed to
subst, which would _re-parse_ the variable reference. This is
unnecessary work.

This modifies the tokenizer to return more metadata about the context in
which it found each string, so that variable references can be handled
directly instead of re-parsed.

This appears to make things somewhat smaller and significantly faster,
though currently this is at the cost of losing "subst."